% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Data}
\format{
\subsection{\code{diabetes}}{

A data frame with 253,777 rows and 6 columns:
\describe{
\item{Person}{ID for each person}
\item{Image}{ID for each image}
\item{x,y}{The cooridinates of each cell}
\item{cellType}{Classified cell type}
\item{Stage}{Stage of the patient}
...
}
}
}
\source{
Currently, \url{https://www.bioconductor.org/packages/release/bioc/html/spicyR.html}

Consider use of \url{https://data.mendeley.com/datasets/cydmwsfztj} or
\url{https://rdrr.io/github/BodenmillerGroup/imcdatasets/man/Damond_2019_Pancreas.html}.
}
\usage{
diabetes
}
\description{
A subset of mass cytometry data from diabetes patients.
}
\details{
Upcoming: Get more complete data set
}
\keyword{datasets}
