% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKFunction.R
\name{.alignKr}
\alias{.alignKr}
\title{Align K and r matrices}
\usage{
.alignKr(K, newK, r, newr)
}
\arguments{
\item{K}{Data.frame of (potentially) columns of K functions for each previous
unit (i.e. the previous K matrix). These are evaluated at the values in r.}

\item{r}{Data.frame of one column for the evaluated r values in K (i.e. the
previous r matrix).}

\item{newr}{Data.frame of one column for the evaluated r values in newK.}

\item{newKData.frame}{of column of K functions for current unit, evaluated
at the values in newr.}
}
\value{
list of three data.frames
\enumerate{
\item K: K functions evaluated at the (potentially) new r values
\item newK: newK evaluated at the (potentially) new r values
\item r: new master vector of evaluated radius r values
}
}
\description{
This (internal) function makes a master r vector and ensures all K functions
are defined at each r.
}
\examples{
# See code for getKFunction This is not an outward function so won't be
#     viewable.
}
