% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.checkData}
\alias{.checkData}
\title{Check Data}
\usage{
.checkData(alignmentMethod)
}
\arguments{
\item{alignmentMethod}{String (or vector of strings) indicating the method of
aligning the variables}
}
\value{
NULL. This function will simply throw an error if there are any
issues.
}
\description{
This (internal) function is used to check data and report an error if any
problems will occur rather than waiting for the function to partially
complete.
}
\details{
Upcoming: Add more data checks into this function
}
\examples{
# See code for computeRandomForest_CVPC. This is not an outward function so
#     won't be viewable.
}
