% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyRandomForest.R
\name{.computeModelAccuracy}
\alias{.computeModelAccuracy}
\title{Compute Model Accuracy}
\usage{
.computeModelAccuracy(oobAcc, outcomes, rGuessSims = 500)
}
\arguments{
\item{oobAcc}{Vector of OOB for each iteration}

\item{outcomes}{Vector of outcomes for the data modeled}

\item{rGuessSims}{Integer indicating the number of iterations used in guess
method}
}
\value{
Data.frame with OOB, guess, and bias accuracy estimates
}
\description{
This (internal) function compute model accuracy based on results from CV.
}
