% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_PC.R
\name{.computeTotalVarImportance}
\alias{.computeTotalVarImportance}
\title{Compute Total Variable Importances}
\usage{
.computeTotalVarImportance(tree, totVarImportance)
}
\arguments{
\item{tree}{Model from rpart for fitting a CART tree.}

\item{totVarImportance}{Data.frame indicating the total (cumulative) variable
importance.}
}
\value{
Data.frame which is the updated totVarImportance adding in the new
tree information.
}
\description{
This (internal) function computes the variable importance for a tree model
and combines it with totVarImportance data.frame.
}
\examples{
# See code for .computeRandomForest_PC. This is not an outward function so won't be
#     viewable.
}
