% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRandomForest.R
\name{.computeTotalVarImportance1}
\alias{.computeTotalVarImportance1}
\title{Compute Total Variable Importances}
\usage{
.computeTotalVarImportance1(tree, totVarImportance = NULL)
}
\arguments{
\item{tree}{Model from rpart for fitting a CART tree.}

\item{totVarImportance}{Data.frame indicating the total (cumulative) variable
importance.}
}
\value{
Data.frame which is the updated totVarImportance adding in the new
tree information.
Compute Total Variable Importances

This (internal) function computes the variable importance for a tree model
and combines it with totVarImportance data.frame.

See use in PCRandomForest.

Data.frame which is the updated totVarImportance adding in the new
tree information.
group'=.getUnderlyingVariable(rownames(tmp1),returnUnique = F))
}
\description{
This (internal) function computes the variable importance for a tree model
and combines it with totVarImportance data.frame.
}
\details{
See use in PCRandomForest.
}
