% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convertList2Dataframe}
\alias{.convertList2Dataframe}
\title{Convert a list to data.frame}
\usage{
.convertList2Dataframe(data_list, typeBind = c("row", "col"), na.omit = F)
}
\arguments{
\item{data_list}{List of data.frames to be combined}

\item{typeBind}{String of row or col indicating if the lists should be
combined using the rows or columns.}

\item{na.omit}{(Optional) Boolean that drops with any NAs if TRUE. The
default is FALSE.}
}
\value{
Data.frame with the data from the list.
}
\description{
This (internal) function converts a list into a data.frame, either by row or
column binding. Some error handling exists, but care should be taken.
}
\details{
See usage in .getPCs or simulatePP.

TODO:: See if this is used in latest refactor
}
