% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCSRData.R
\name{.generateCSRData}
\alias{.generateCSRData}
\title{Generate Completely Spatially Random Data (Point Process)}
\usage{
.generateCSRData(
  xRange = c(0, 1),
  yRange = c(0, 1),
  kappa = 25,
  requireOne = T,
  cellType = "A"
)
}
\arguments{
\item{xRange}{(Optional) Two value numeric vector indicating the size of the
region in the x-direction. Default is c(0,1).}

\item{yRange}{(Optional) Two value numeric vector indicating the size of the
region in the y-direction. Default is c(0,1).}

\item{kappa}{(Optional) Numeric influencing the number of points. Kappa times
area defines the mean of a Poisson random variable which is randomly
drawn to indicate the number of cells in the area. Default is 25.}

\item{requireOne}{(Optional) Boolean that indicating if at least one point
must be placed. This can affect the Poisson distribution, but is valuable
in ensuring a non-empty region. Default is TRUE.}

\item{cellType}{(Optional) Value (often character) indicating the label the
points in the process should be given. Default is A.}

\item{percentAwayFromEdge}{(Optional) Numeric in \link{0,0.5} which limits how
close to the edge a point may be placed. This can be used to limit or
test edge effects. Default is 0.025 on each edge.}
}
\value{
Data.frame with x, y, and cellType specified. Each row is a new
point.
}
\description{
This (internal) function generating a completely spatially random (CSR)
process on some region. Some Poisson number of points are placed
uniformly in the region by both x and y.
}
\details{
See use in simulatePP_Mult_PCA_Meta.
}
