% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{.generateCSRPatterns}
\alias{.generateCSRPatterns}
\title{Generate CSR Point Patterns}
\usage{
.generateCSRPatterns(
  stageName,
  peoplePerStage,
  imagesPerPerson,
  kappas,
  cellTypes,
  kappaSep = T,
  imageAdj = 0,
  personAdj = 0
)
}
\arguments{
\item{stageName}{String inidicating the outcome name that should be given.}

\item{peoplePerStage}{Numeric indicating the number of units per
outcome.}

\item{imagesPerPerson}{Numeric indicating the number of repeated
measures.}

\item{kappas}{Numeric vector directing the number of agents per type per
image.}

\item{cellTypes}{Vector with the agent types of interest being generated.
Corresponds to the kappas.}

\item{kappaSep}{(Optional) Boolean. Current use understood but not used. May
remove.}
}
\value{
Data.frame containing each point in the defined CSR patterns.

The data.frame has columns for outcome, x coordinate, y coordinate, agent
type, unit, and unique repeated measure id.
}
\description{
This (internal) function is used to generate CSR point patterns. It calls
.generateCSRData while also adding necessary data and allowing vectors to
be entered into the function.
}
\details{
See usage in simulatePP.

Upcoming: Try to eliminate or indicate when kappaSep is used.
}
