% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.generateCVVariableImportancePlot}
\alias{.generateCVVariableImportancePlot}
\title{Generate Variable Importance Plots}
\usage{
.generateCVVariableImportancePlot(
  giniData,
  viData,
  accData,
  noiseMap,
  KFunctions,
  metaNames,
  alpha,
  alignmentMethod = "Mult",
  curveData = NULL,
  subsetPlotSize = 50
)
}
\arguments{
\item{giniData}{Data.frame organizing the gini index metrics of the
variables. Columns of var, avg, sd, lower, and upper.}

\item{viData}{Data.frame organizing the variable importance index metrics of
the variables. Columns of var, avg, sd, lower, and upper.}

\item{accData}{Data.frame}

\item{noiseMap}{Data.frame for matching noise variables to noise groups. The
columns are noiseVar (the synthetic interactions of which there are
syntheticMetas), dataVar (name of noise group), and synType ('Meta',
which will compare to 'K' from other functions).}

\item{KFunctions}{Vector of the underlying K functions of interest. These
will be the ones plotted.}

\item{metaNames}{Vector of the underlying Meta-variables of interest. These
will be the ones plotted.}

\item{alpha}{Numeric in (0,1) indicating the significance for the cutoff.}

\item{alignmentMethod}{(Optional) String indicating the method of aligning
variables via noise variables. The options are 'Add', 'Mult', or c('Add',
'Mult'). The default value is 'Mult'.}

\item{curveData}{(Optional) Data.frame with two columns (gini, vi) and rows
equal to the number of variables being considered. This defines the
values used to create the curved cutoff. Default is NULL.}

\item{subsetPlotSize}{(Optional) Numeric indicating the number of top
variables to include in a subset graph (note if there are less variables)
than this value indicates then no subset graph will be produced. Default
is 50.}
}
\value{
List containing three plots.
1. varImpPlot: giniPlot and viPlot together with cross-validated error.
2. viPlot: plot with each K function and Meta-variable aligned with
noise variable cutoff using variable importance metric.
3. giniPlot: plot with each K function and Meta-variable aligned with
noise variable cutoff using gini metric.
}
\description{
This (internal) function generates the CV variable importance plots for the
RandomForest_CVPC function. It add CV errors and manages error alignment.
}
\details{
Upcoming: Remove Gini.
}
\examples{
# See code for computeRandomForest_CVPC. This is not an outward function so
#     won't be viewable.
}
