% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.generateKNoiseGroup}
\alias{.generateKNoiseGroup}
\title{Generate K Noise Group}
\usage{
.generateKNoiseGroup(
  syntheticKs,
  cellData,
  kappaL,
  kappaR,
  agentName,
  outcome,
  unit,
  repeatedId
)
}
\arguments{
\item{syntheticKs}{Numeric indicating the number of variables
in the K noise groups. If the value is 0, no K noise variables are
generated.}

\item{cellData}{Data.frame indicating the cells used to create data. It
contains outcome, unit, (possibly) repeatedId, and agentType.}

\item{kappaL}{Numeric value for intensity of left agent (A in traditional A_A
writing of interaction terms).}

\item{kappaR}{Numeric value for intensity of right agent (B in traditional
A_B writing of interaction terms).}

\item{agentName}{String indicating the K function it is matched with. This
may be a general term.}

\item{outcome}{String indicating the column name with the outcome in
cellData.}

\item{unit}{String indicating the column name with the unit in cellData.}

\item{repeatedId}{String indicating the column name with the unique id from
repeated measures in cellData.}
}
\value{
Data.frame for all the cells in the noise group. This will contain
columns of x, y, and cellType.
}
\description{
This (internal) function generates noise groups of syntheticKs patterns.
}
\examples{
# See code for .generateSyntheticKs. This is not an outward function so won't
#     be viewable.
}
