% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.generateMetaNoiseGroup}
\alias{.generateMetaNoiseGroup}
\title{Generate Meta Noise Group}
\usage{
.generateMetaNoiseGroup(data, meta, syntheticMetas)
}
\arguments{
\item{data}{Data.frame of outcome and predictors (PCs and meta-variables).
Note, currently Unit or repeated measures should not be included.
Generally use the results from getPCAData, potentially with meta-
variables attached. Really only column with meta column needed}

\item{meta}{String of the column name indicating the meta variable to be
generated.}

\item{syntheticMetas}{Numeric indicating the number of synthetic meta
variables in each noise group.}
}
\value{
Data.frame with syntheticMetas columns and observations (rows) of
data. Each column is a bootstrapped sample of the meta variable.
}
\description{
This (internal) function that will bootstrap meta variables to create
synthetic variables.
}
\examples{
# See code for .generateSyntheticMetas. This is not an outward function so
#     won't be viewable.
}
