% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCurveVI.R
\name{.generateNoiseCurve}
\alias{.generateNoiseCurve}
\title{Generate Noise Curve}
\usage{
.generateNoiseCurve(
  dat,
  nSims,
  outcome,
  unit,
  repeatedId,
  noiseMap,
  KFunctions,
  metaNames,
  syntheticMetaNames,
  alpha = 0.05,
  silent = F,
  alignmentMethod = "Add"
)
}
\arguments{
\item{dat}{Data.frame used to fit random forest. See data given to
computeRandomForest_CVPC, but note synthetics will also be attached.}

\item{nSims}{Numeric indicating the number of simulations to bootstrap the
data .}

\item{outcome}{String indicating the column name with the outcome in dat.}

\item{unit}{String indicating the column name with the unit in dat.}

\item{repeatedId}{String indicating the column name with the unique id from
repeated measures in dat.}

\item{noiseMap}{Data.frame for matching noise variables to noise groups. The
columns are noiseVar (the synthetic interactions of which there are
syntheticKs), dataVar (name of noise group), and synType ('K', which will
compare to 'Meta' from other functions).}

\item{KFunctions}{Vector indicating the K functions to be considered. These
string names should be in the column names of dat.}

\item{metaNames}{Vector indicating the meta-variables to be considered. These
string names should be in the column names of dat.}

\item{syntheticMetaNames}{Vector indicating the synthetic meta-variables to
be considered. These string names should be in the column names of dat.}

\item{alpha}{(Optional) Numeric in (0,1) indicating the significance used
throughout the analysis. Default is 0.05.}

\item{silent}{(Optional) Boolean indicating if output should be suppressed
when the function is running. Default is FALSE.}

\item{alignmentMethod}{(Optional) String indicating the method of aligning
the variables. The options are 'Add', 'Mult', or c('Add','Mult'). The
default is 'Add'.}
}
\value{
Data.frame with 2 - 4 columns. If only one alignmentMethod is given,
a 2 column data.frame is returned with the first column for gini and the
second column for variable importance (both standardized). If two
alignmentMethod values are given, the first two columns related to gini
and variable importance for the additional standardization while columns
3-4  related to the multiplication standardization.
}
\description{
This (internal) function generates the noise course of data. It indicates
the variable importance expected for a group of comparable noise
variables. This can combat mistakenly thinking the variable with the
greatest variable importance is significant.
}
\examples{
# See code for computeRandomForest_CVPC. This is not an outward
#     function so won't be viewable.
}
