% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.generateNoiseVariable}
\alias{.generateNoiseVariable}
\title{Generate Noise Variable}
\usage{
.generateNoiseVariable(
  syntheticKs,
  kappaL,
  kappaR,
  currentInfo,
  agentName,
  outcome,
  unit,
  repeatedId
)
}
\arguments{
\item{syntheticKs}{Numeric indicating the number of variables
in the K noise groups. If the value is 0, no K noise variables are
generated.}

\item{kappaL}{Numeric value for intensity of left agent (A in traditional A_A
writing of interaction terms).}

\item{kappaR}{Numeric value for intensity of right agent (B in traditional
A_B writing of interaction terms).}

\item{currentInfo}{Data.frame with the information for the patterns. Includes
outcome, unit, and (potentially) repeatedId values.}

\item{agentName}{String indicating the K function it is matched with. This
may be a general term.}

\item{outcome}{String indicating the column name with the outcome in
cellData.}

\item{unit}{String indicating the column name with the unit in cellData.}

\item{repeatedId}{String indicating the column name with the unique id from
repeated measures in cellData.}
}
\value{
Data.frame with cells for noise variables. The columns are x, y, and
cellType.
}
\description{
This (internal) function generates the cells for the syntheticKs K functions
--two patterns for each side.
}
\examples{
# See code for .generateKNoiseGroup. This is not an outward function so won't
#     be viewable.
}
