% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.generateSyntheticKs}
\alias{.generateSyntheticKs}
\title{Generate Synthetic K functions}
\usage{
.generateSyntheticKs(
  syntheticKs,
  data,
  cellData,
  outcome,
  unit,
  repeatedId,
  metaNames,
  generalSyntheticK,
  silent = F
)
}
\arguments{
\item{syntheticKs}{Numeric indicating the number of variables
in the K noise groups. If the value is 0, no K noise variables are
generated.}

\item{data}{Data.frame of outcome and predictors (PCs and meta-variables).
Note, currently Unit or repeated measures should not be included.
Generally use the results from getPCAData, potentially with meta-
variables attached.}

\item{cellData}{Data.frame indicating the cells used to create data. It
contains outcome, unit, (possibly) repeatedId, agentType.}

\item{outcome}{String indicating the column name with the outcome in both
data and cellData.}

\item{unit}{String indicating the column name with the unit in both data and
cellData.}

\item{repeatedId}{String indicating the column name with the unique id from
repeated measures in both data and cellData.}

\item{metaNames}{Vector indicating the meta-variables to be considered.}

\item{generalSyntheticK}{Boolean indicating if a general K noise group
should be used or specialized K noise groups should be used.}

\item{silent}{(Optional) Boolean indicating if output should be suppressed
when the function is running. Default is FALSE.}
}
\value{
List with two elements:
1. noiseMap: Data.frame for matching noise variables to noise groups.
The columns are noiseVar (the synthetic interactions
of which there are syntheticKs), dataVar (name of
noise group), and synType ('K', which will compare
to 'Meta' from other functions).
2. pcaData: Data.frame with nPCs equal to that in data. This contains
the pcaData for the noise variables. See values
returns from getPCAData, but outcome, person,
repeatedId, and PCs for synthetic Ks.
}
\description{
This (internal) function generates the synthetic K functions (in PCs).
}
\details{
Upcoming: See following experiments:
- Test if Kappa should be related to data or not
- Determine if window size c(1,1) is reasonable for all situation
}
\examples{
# See code for computeRandomForest_CVPC. This is not an outward function so
#     won't be viewable.
}
