% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyRandomForest.R
\name{.generateVIPlot}
\alias{.generateVIPlot}
\title{Generate VI Plot}
\usage{
.generateVIPlot(
  viData,
  accData,
  NoiseCutoff,
  InterpolationCutoff,
  subsetPlotSize
)
}
\arguments{
\item{viData}{Data.frame with var, est, and sd}

\item{accData}{Data.frame with oob, bias, and guess estimates of model
accuracy}

\item{NoiseCutoff}{Numeric indicating the vertical noise cutoff}

\item{InterpolationCutoff}{Vector of numerics for the curved cutoff}

\item{subsetPlotSize}{Integer indicating number of interactions to have on
the subset plot. Note if there are fewer in the base, no subset plot will
be created}
}
\value{
List with VI figures (one or two)
}
\description{
This (internal) function is used in creation of the VI plot.
}
