% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCurveVI.R
\name{.getCurveValues}
\alias{.getCurveValues}
\title{Get Curve Values}
\usage{
.getCurveValues(stdData, alpha)
}
\arguments{
\item{stdData}{Data.frame with the variables as rows and the point estimates
for each run, given in each column.}

\item{alpha}{Numeric in (0,1) for finding the 1-alpha quantile.}
}
\value{
This function returns a vector of numerics indicating the expected
noise values for an equivalent number of random variables.
}
\description{
The (internal) function gets the 1-alpha quantile of each relative location
in data. This accounts for the fact that variables may change relative
locations on different runs.
}
\examples{
# See code for .generateNoiseCurve. This is not an outward
#     function so won't be viewable.
}
