% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getFolds}
\alias{.getFolds}
\title{Create K folds}
\usage{
.getFolds(x, K)
}
\arguments{
\item{x}{Vector of data (or often indices)}

\item{chunksN}{Numeric for the number of chunks, or folds, desired}
}
\value{
a list with length chunksN, with each containing approximately the
same number of points
}
\description{
This (internal) function creates K folds of x. Note these folds are scrambled
data, but then ordered in each fold.
}
\examples{
.getFolds(1:10,3)
}
