% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.getIndividualCutoffs}
\alias{.getIndividualCutoffs}
\title{Get Individual Cutoffs}
\usage{
.getIndividualCutoffs(noiseMap, giniData, viData, alignmentMethod, alpha)
}
\arguments{
\item{noiseMap}{Data.frame for matching noise variables to noise groups. The
columns are noiseVar (the synthetic interactions of which there are
syntheticMetas), dataVar (name of noise group), and synType ('Meta',
which will compare to 'K' from other functions).}

\item{giniData}{Data.frame organizing the gini index metrics of the
variables. Columns of var, avg, sd, lower, and upper.}

\item{viData}{Data.frame organizing the variable importance index metrics of
the variables. Columns of var, avg, sd, lower, and upper.}

\item{alignmentMethod}{String indicating the method of aligning variables via
noise variables. The options are 'Add' or 'Mult'.}

\item{alpha}{Numeric in (0,1) indicating the significance for the cutoff.}
}
\value{
List of two data.frames. Each data.frame has columns dataVar
(variable names), synType (Synthetic types used for standardization),
cutoff (the cutoff for the variable), and adj (the aligned variables).
1. giniCutoff_df: data.frame based on gini index.
2. viCutoff_df: data.frame based on variable importance metric.
}
\description{
This (internal) function gets the 1-alpha cutoffs for the variables of
interest based on noise variables mapped via the noiseMap. Cutoffs for
Gini and variable importance are individual considered. Alignments of the
cutoffs are done either using multiplication or addition.
}
\examples{
# See code for .generateCVVariableImportancePlot. This is not an outward
#     function so won't be viewable.
}
