% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getMode}
\alias{.getMode}
\title{Get Mode of vector}
\usage{
.getMode(x)
}
\arguments{
\item{x}{Vector of some items}
}
\value{
The mode(s) of the vector
}
\description{
Get Mode of vector
}
\examples{
.getMode(c(1,2,3,1))
.getMode(c('A','B','C','A','B'))
.getMode(c(1,2,3,'A','A'))
.getMode(c(1,2,3,'A','A',2,2,2))
}
