% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCAData.R
\name{.getPCs}
\alias{.getPCs}
\title{Get Principal Components}
\usage{
.getPCs(rKData, agents, nPCs, nbasis = 21, silent = F)
}
\arguments{
\item{rKData}{data.frame with the first column being the checked radius and the
rest relating to the K function for each unit at those points. NA columns
for any K functions that could not be computed will be handled.}

\item{agents}{Two value vector indicating the two agents used for the K
function, the first to the second.}

\item{nPCs}{Numeric indicating the number of principal components to compute}

\item{nbasis}{Numeric indicating the number of basis functions to use to fit
the data to a bspline basis.}
}
\value{
Data.frame with the outcomes, units, then principal component scores.
}
\description{
This function coverts the radius and K functions into principal components
and recturns the scores.
}
\details{
See getPCAData for use.
}
