% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCurveVI.R
\name{.getStdCurveData}
\alias{.getStdCurveData}
\title{Get standardized Curve Data}
\usage{
.getStdCurveData(
  avgGini,
  avgVI,
  alpha,
  KFunctions,
  metaNames,
  noiseMap,
  alignmentMethod
)
}
\arguments{
\item{avgGini}{Data.frame with gini index values from a random forest in
columns, each column representing a trial and a row for each variable,
with the first column being a variable name column.}

\item{avgVI}{Data.frame with variable importance values from a random forest
in columns, each column representing a trial and a row for each variable,
with the first column being a variable name column.}

\item{alpha}{Numeric in (0,1) indicating the significance used
throughout the analysis. Default is 0.05.}

\item{KFunctions}{Vector indicating the K functions of interest.}

\item{metaNames}{Vector indicating the meta-variables of interest.}

\item{noiseMap}{Data.frame for matching noise variables to noise groups.
The columns are noiseVar (the synthetic interactions of which
there are syntheticKs), dataVar (name of noise group), and
synType ('K', which will compare to 'Meta' from other
functions).}

\item{alignmentMethod}{String indicating the method of aligning the
variables. The options are 'Add' or 'Mult'.}
}
\value{
List containing two data.frame elements of the same structure. The
first column is var and indicates the variable and subsequent columns are
the standardized values for each sim. The list elements are:
\enumerate{
\item gini: Data.frame built using the gini index.
\item vi: Data.frame built using the variable importance metric.
}
}
\description{
This (internal) function gets variable importance metrics and standardizes
the data used in the curved cutoff.
}
\examples{
# See code for .generateNoiseCurve. This is not an outward
#     function so won't be viewable.
}
