% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.getUnderlyingNamesAndPCsCount}
\alias{.getUnderlyingNamesAndPCsCount}
\title{Get Underlying Names and Principal Components Counts}
\usage{
.getUnderlyingNamesAndPCsCount(namesOfInterest)
}
\arguments{
\item{namesOfInterest}{Vector of variable names, names with _PC# with them}
}
\value{
List with two elements:
\enumerate{
\item Vars: Underlying function names
\item PCs: Numeric indicating the number of principal components
}
}
\description{
This (internal) function get the underlying functions and the total principal
components from the variable function names.
}
\examples{
# See code for .generateSyntheticKs. This is not an outward function so won't
#     be viewable.
}
