% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mergeListsToDF}
\alias{.mergeListsToDF}
\title{Merge Lists into Base Data.frame}
\usage{
.mergeListsToDF(df, lists, dfCol, listsDFCol)
}
\arguments{
\item{df}{Data.frame with a column named by baseCol}

\item{lists}{Lists of data.frames with each containing listCol column
matching the values in df$baseCol.}

\item{dfCol}{String with column name for matching column in df}

\item{listsDFCol}{String with column name for matching column in lists}
}
\value{
Data.frame df with values from lists appended
}
\description{
This (internal) function appends items in list to a data.frame by the given
column names.
}
\details{
See usage in getPCAData.

TODO:: Verify use in latest refactor
}
