% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyModel.R
\name{.permuteData}
\alias{.permuteData}
\title{Generate Permuted Data}
\usage{
.permuteData(
  data_base,
  outcome,
  unit,
  synthetics,
  KFunctions,
  metaNames,
  underlyingDataAlignedFunctions,
  nPCs,
  attach.data = FALSE,
  permute.data = FALSE
)
}
\arguments{
\item{data_base}{Data.frame with outcome, unit, and some variables}

\item{outcome}{(Optional) String indicating the outcome column name in data.
Default is the first column of data.}

\item{unit}{(Optional) String indicating the unit column name in data.
Default is the second column of data.}

\item{synthetics}{(Optional)}

\item{KFunctions}{Vector of the K Function names.}

\item{metaNames}{(Optional) Vector indicating the meta-variables to be
considered. Default is NULL.}

\item{underlyingDataAlignedFunctions}{Vector of the column names aligned to
underlying function. Generated in funkyModel.}

\item{nPCs}{Integer indicating the number of principle components}

\item{attach.data}{(Optional) Boolean indicating if true data should be
attached. Default is FALSE.}

\item{permute.data}{(Optional) Boolean indicating if true data should be
permuted. Note this will only be valuable if attach.data is TRUE. Default is
FALSE.}
}
\value{
List with two entries:
\enumerate{
\item data: Data.frame with outcome, unit, and data (may be only permutate or real
and permuted based on settings)
\item metaNames: Vector of the metaNames in the data
}
}
\description{
This (internal) function permutes the data for use in model via noise or variable
permutation
}
