% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{.placeClusteredPts}
\alias{.placeClusteredPts}
\title{Place Clustered Points}
\usage{
.placeClusteredPts(
  currXY,
  cell,
  numPts,
  varValue,
  xRange = c(0, 1),
  yRange = c(0, 1)
)
}
\arguments{
\item{currXY}{Vector of two numerics, the first relating the x coordinate
and the second to the y coordinate. These turn into the mean of the
normal distribution}

\item{cell}{String indicating the agent type of the currently placed agents.}

\item{numPts}{Numeric indicating the number of agents to place.}

\item{varValue}{Numeric giving the variance of the normal distribution for
placement of the agents.}

\item{xRange}{(Optional) Vector of two values indicating the x range of the
region. Default is c(0,1).}

\item{yRange}{(Optional) Vector of two values indicating the x range of the
region. Default is c(0,1).}
}
\value{
Data.frame with placed cells.

The data.frame has 3 columns, x, y, and cellType.
}
\description{
This (internal) function places points following a normal distribution around
some given corridinates.
}
\details{
See usage in simulatePP.

Upcoming: Generalize naming
}
