% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.plotCVVariableImportance}
\alias{.plotCVVariableImportance}
\title{Plot CV Variable Importance}
\usage{
.plotCVVariableImportance(underlyingData, ylabel, cutoff, curveDat)
}
\arguments{
\item{underlyingData}{Data.frame with 4 columns (var, avgVar, lower, upper).
See the results from .standardizeVarImpData.}

\item{ylabel}{String for the y-axis label}

\item{cutoff}{Numeric in \link{0,1} for the red noise line}

\item{curveDat}{(Optional) Vector of numerics of length nrow(underlyingData).
This will add the curved line as developed in .generateNoiseCurve.}
}
\value{
ggplot2 plot for the CV variable importance plot
}
\description{
This (internal) function produces the CV variable importance plots with the
noise cutoff and CV intervals.
}
\details{
Upcoming: Rename ylabel (as we flip cooridinates its actually x-axis we see)
}
\examples{
# See code for .generateCVVariableImportancePlot. This is not an outward
#     function so won't be viewable.
}
