% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_PC.R
\name{.plotVariableImportance}
\alias{.plotVariableImportance}
\title{Plot Variable Importance (Ensure Data)}
\usage{
.plotVariableImportance(varImportanceData = NULL, model = NULL)
}
\arguments{
\item{varImportanceData}{(Optional) Data.frame for the variable importance
information.}

\item{model}{(Optional) Random forest model from .computeRandomForest_PC.}
}
\value{
grid.arrange containing two ggplots
}
\description{
This (internal) function organizes the data to plot the variable importance
for a .computeRandomForest_PC model. One of the inputs is necessary.
}
\details{
Upcoming: See TEST comment
}
\examples{
# See code for .computeRandomForest_PC. This is not an outward function so won't be
#     viewable.
}
