% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyRandomForest.R
\name{.plotVI}
\alias{.plotVI}
\title{Plot VI}
\usage{
.plotVI(viData, accData, NoiseCutoff, InterpolationCutoff)
}
\arguments{
\item{viData}{Data.frame with var, est, and sd}

\item{accData}{Data.frame with oob, bias, and guess estimates of model
accuracy}

\item{NoiseCutoff}{Numeric indicating the vertical noise cutoff}

\item{InterpolationCutoff}{Vector of numerics for the curved cutoff}
}
\value{
ggplot2 figure object for the VI plot
}
\description{
This (internal) function standardizing and plots the variable importance
values.
}
