% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKFunction.R
\name{.rK2DF}
\alias{.rK2DF}
\title{Convert r and K to a Data.frame}
\usage{
.rK2DF(K_list, r_list)
}
\arguments{
\item{K_list}{List of data.frames from each unit. Each data.frame has the
K function for each unit evaluated at the corresponding r in r_list.}

\item{r_list}{List of data.frames from each unit. Each data.frame has the
evaluted r values for each unit used in the corresponding K in K_list.}
}
\value{
Data.frame with the first column being the evaluated r and the rest
being the evaluted K functions for each unit.
}
\description{
This converts the r and K matrices from different units into a single
data.frame.
}
\examples{
# See code for getKFunction This is not an outward function so won't be
#     viewable.
}
