% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.specify_decimal}
\alias{.specify_decimal}
\title{Specify Decimals}
\usage{
.specify_decimal(x, k)
}
\arguments{
\item{x}{Numeric to display k decimals of.}

\item{k}{Numeric indicating the number of decimals to display}
}
\value{
String with numeric x to k decimal places
}
\description{
This (internal) function ensures a numeric displays a given number of
decimals. This is different from "round" in that it will add zeros as
necessary.
}
\examples{
.specify_decimal(10.123,1)
.specify_decimal(10.123,3)
.specify_decimal(10.123,5)
}
