% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_CVPC.R
\name{.standardizeVarImpData}
\alias{.standardizeVarImpData}
\title{Standardize Variable Importance Data}
\usage{
.standardizeVarImpData(
  giniData,
  viData,
  KFunctions,
  metaNames,
  giniCutoff_df,
  viCutoff_df,
  alignmentMethod
)
}
\arguments{
\item{giniData}{Data.frame organizing the gini index metrics of the
variables. Columns of var, avg, sd, lower, and upper.}

\item{viData}{Data.frame organizing the variable importance index metrics of
the variables. Columns of var, avg, sd, lower, and upper.}

\item{KFunctions}{Vector of the underlying K functions of interest. These
will be the ones plotted.}

\item{metaNames}{Vector of the underlying Meta-variables of interest. These
will be the ones plotted.}

\item{giniCutoff_df}{Data.frame built from gini index having columns dataVar
(variable names), synType (Synthetic types used for standardization),
cutoff (the cutoff for the variable), and adj (the aligned variables).}

\item{viCutoff_df}{Data.frame built from variable importance metric having
columns dataVar (variable names), synType (Synthetic types used for
standardization), cutoff (the cutoff for the variable), and adj (the
aligned variables).}

\item{alignmentMethod}{String indicating the method of aligning variables via
noise variables. The options are 'Add' or 'Mult'.}
}
\value{
List containing two named data.frame elements.
1. stdGini: A data.frame with 4 columns (var, avgVal, lower, and upper).
The column var indicates the variable name (from KFunctions
and metaNames), the column avgVal indicates the avg from
giniData aligned using alignmentMethod. Columns lower and
upper are the aligned 1-alpha confidence intervals from
giniData.
2. stdVI: A data.frame with 4 columns (var, avgVal, lower, and upper).
The column var indicates the variable name (from KFunctions
and metaNames), the column avgVal indicates the avg from
viData aligned using alignmentMethod. Columns lower and
upper are the aligned 1-alpha confidence intervals from
viData.
}
\description{
This (internal) function returns the variables (specified in KFunctions and
metaNames), standardized with lower and upper 1-alpha confidence
intervals, respective to gini index and variable importance measures and
aligned as specified in alignmentMethod.
}
\examples{
# See code for .generateCVVariableImportancePlot. This is not an outward
#     function so won't be viewable.
}
