% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCurveVI.R
\name{.stdCurveData}
\alias{.stdCurveData}
\title{Standardize Curve Data}
\usage{
.stdCurveData(datAvg, cutoff_df, KFunctions, metaNames, alignmentMethod)
}
\arguments{
\item{datAvg}{Data.frame with first column, var, indicating variable (inc.
noise variables) and subsequent columns indicating the values for
importance of each variable.}

\item{cutoff_df}{Data.frame with 4 columns. The column dataVar indicates the
variables which are related, synType indicates if the variable is K or
meta-variable, adj indicates how much the related variables will need to
be adjusted, and cutoff indicates the current cutoffs for the variables.}

\item{KFunctions}{Vector indicating the K functions of interest.}

\item{metaNames}{Vector indicating the meta-variables of interest.}

\item{alignmentMethod}{String indicating the method of aligning the
variables. The options are 'Add' or 'Mult'.}
}
\value{
Data.frame with the first column (var) indicating variables of
interest and the rest being the standardized values from datAvg.
}
\description{
This (internal) function standardizes the data in datAvg using cutoffs in
cutoff_df. It returns the standardization for variables of interest,
indicated in KFunctions and metaNames. Standardization is completed
according to alignmentMethod.
}
\examples{
# See code for .getStdCurveData. This is not an outward
#     function so won't be viewable.
}
