% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyRandomForest.R
\name{.summData}
\alias{.summData}
\title{Get Data Summary}
\usage{
.summData(dat)
}
\arguments{
\item{dat}{Data.frame with columns for different iterations, rows are the
variables.}
}
\value{
Data.frame with var, est, and sd as columns
}
\description{
This is an internal functional for summarizing data from different iterations.
}
