% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkyRandomForest.R
\name{funkyRandomForest}
\alias{funkyRandomForest}
\title{Fit a Random Forest model with PC data (Using CV for Improvements)}
\usage{
funkyRandomForest(
  data,
  K = 10,
  outcome = colnames(data)[1],
  unit = colnames(data)[2],
  metaNames = NULL,
  synthetics = 100,
  alpha = 0.05,
  silent = F,
  rGuessSims = 500,
  subsetPlotSize = 25,
  nTrees = 500,
  method = "class"
)
}
\arguments{
\item{data}{Data.frame of outcome and predictors (PCs and meta-variables).
Generally use the results from getPCAData, potentially with meta-
variables attached.}

\item{K}{(Optional) Numeric indicating the number of folds to use in K-fold
CV. The default is 10.}

\item{outcome}{(Optional) String indicating the column name with the outcome
in both data and cellData. Default is the first column of data.}

\item{unit}{(Optional) String indicating the column name with the unit in
both data and cellData. Default is the second column of data.}

\item{metaNames}{(Optional) Vector indicating the meta-variables to be
considered. Default is NULL.}

\item{alpha}{(Optional) Numeric in (0,1) indicating the significance used
throughout the analysis. Default is 0.05.

This is used in selection of noise variables and in building the CV
intervals.}

\item{silent}{(Optional) Boolean indicating if output should be suppressed
when the function is running. Default is FALSE.}

\item{rGuessSims}{(Optional) Numeric value indicating the number of
simulations used for guessing and creating the guess estimate on the
plot. Default is 500.}

\item{subsetPlotSize}{(Optional) Numeric indicating the number of top
variables to include in a subset graph (note if there are less variables)
than this value indicates then no subset graph will be produced. Default
is 25.}

\item{nTrees}{(Optional) Numeric indicating the number of trees in each
forest. The default is 500.}

\item{repeatedId}{(Optional) String indicating the column name with the
unique id from repeated measures in both data and cellData. Default is
NULL.}

\item{cellData}{(Optional) Data.frame indicating the cells used to create
data. It contains outcome, unit, (possibly) repeatedId, agentType.}

\item{syntheticKs}{(Optional) Numeric indicating the number of variables
in the K noise groups. If the value is 0, no K noise variables are
generated. Default is 100.}

\item{syntheticMetas}{(Optional) Numeric indicating the number of variables
in the meta noise groups. If the value is 0, no meta noise variables are
generated. Default is 100.}

\item{generalSyntheticK}{(Optional) Boolean indicating if a general K noise
group should be used or specialized K noise groups should be used.
Default is TRUE.}

\item{curvedSigSims}{(Optional) Numeric indicating the number of simulations
used to create the curved lines. Default is 100.}

\item{alignmentMethod}{(Optional) String indicating the method of aligning
variables via noise variables. The options are 'Add', 'Mult', or c('Add',
'Mult'). The default value is c('Add','Mult').}
}
\value{
List with the following items:
\enumerate{
\item VariableImportance: Data.frame with the results of variable
importance indices from the models and CV. The columns are
var, est, and sd. The columns lower and upper are made with
significance alpha.
\item AccuracyEstimate: Data.frame with model accuracy estimates:
out-of-bag accuracy (OOB), biased estimate (bias), and
random guess (guess). The columns are OOB, bias, and guess.
\item NoiseCutoff: Numeric indicating noise cutoff (vertical line)
\item InterpolationCutoff: Vector of numerics indicating the
interpolation cutoff (curved line)
\item AdditionalParams: List of additional params for reference:
Alpha and subsetPlotSize.
\item viPlot: ggplot2 object for vi plot with standardized results.
It displays ordered underlying functions and meta-variables
with point estimates, sd, noise cutoff, and interpolation
cutoff all based on variable importance values
\item subset_viPlot: ggplot2 object for vi plot with standardized
results and only top subsetPlotSize variables. It displays
ordered underlying functions and meta-variables with point
estimates, sd, noise cutoff, and interpolation cutoff all
based on variable importance values
}
\description{
The function fits a random forest model to the data along with using cross-
validation to quantify variable importance.
}
\details{
Warning: If there are no sythentics, this may break (will fix it eventually)
}
\examples{
dat <- simulatePP(cellVarData=
                    data.frame('stage'=c(0,1), 'A'=c(0,0), 'B'=c(1/50,1/50)),
                  cellKappaData=data.frame(
                                      'cell'=c('A','B'),
                                      'clusterCell'=c(NA,'A'),
                                      'kappa'=c(20,5)),
                  peoplePerStage=100,
                  imagesPerPerson=1,
                  silent=F )
pcaData <- getPCAData(dat,repeatedUniqueId='Image',
                      xRange = c(0,1),  yRange = c(0,1), silent=F)
pcaMeta <- simulateMeta(pcaData,
                        metaInfo = data.frame(
                           'var'=c('randUnif','randBin','corrNorm'),
                           'rdist'=c('runif','rbinom','rnorm'),
                           'Stage_0'=c('0.5','0.5','1'),
                           'Stage_1'=c('0.5','0.5','2')))
rfcv <- funkyRandomForest(data=pcaMeta,outcome = 'Stage',unit='Person',
                          metaNames=c('randUnif','randBin','corrNorm'))

dat <- simulatePP(cellVarData=
                    data.frame('stage'=c(0,1), 'A'=c(0,0), 'B'=c(1/50,1/100)),
                  cellKappaData=data.frame(
                                     'cell'=c('A','B'),
                                     'clusterCell'=c(NA,'A'),
                                     'kappa'=c(20,5)),
                  peoplePerStage=100,
                  imagesPerPerson=1,
                  silent=F )
pcaData <- getPCAData(dat,repeatedUniqueId='Image',
                      xRange = c(0,1),  yRange = c(0,1), silent=F)
pcaMeta <- simulateMeta(pcaData,
                        metaInfo = data.frame(
                            'var'=c('randUnif','randBin','corrNorm'),
                            'rdist'=c('runif','rbinom','rnorm'),
                            'Stage_0'=c('0.5','0.5','1'),
                            'Stage_1'=c('0.5','0.5','2')))
rfcv <- funkyRandomForest(data=pcaMeta,outcome = 'Stage',unit='Person',
                          metaNames=c('randUnif','randBin','corrNorm'),
                          subsetPlotSize = 2)
}
