% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungcancer.R
\docType{data}
\name{lungcancer}
\alias{lungcancer}
\title{Lung Cancer Data}
\format{
\subsection{\code{lungcancer}}{

A data frame with 79,906 rows and 7 columns:
\describe{
\item{Patient}{ID for each patient}
\item{Image}{ID for each image of a patient}
\item{x,y}{The cooridinates of each cell}
\item{Cancer}{Classified type of cancer}
\item{SmokingStatusLong}{Long term smoking status of the patient}
\item{cellType}{The classified cell type of each cell}
...
}
}
}
\usage{
lungcancer
}
\description{
A subset of mass cytometry data from lung cancer patients.
}
\keyword{datasets}
