% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRandomForest.R
\name{PCRandomForest}
\alias{PCRandomForest}
\title{Compute Random Forest for Data with multiple PC (Along With Meta-Variables)}
\usage{
PCRandomForest(
  data,
  outcome = colnames(data)[1],
  unit = colnames(data)[2],
  repeatedId = NULL,
  nTrees = 500,
  varImpPlot = T,
  metaNames = NULL,
  keepModels = T,
  varSelPercent = 0.8,
  method = "class"
)
}
\arguments{
\item{data}{Data.frame of outcome and predictors (PCs and meta-variables).
Note, currently Unit or repeated measures should not be included.
Generally use the results from getPCAData, potentially with meta-
variables attached.}

\item{outcome}{(Optional) String indicating the outcome column name in data}

\item{nTrees}{(Optional) Numeric indicating the number of trees to use in the
random forest model. Default is 500.}

\item{varImpPlot}{(Optional) Boolean indicating if variable importance plots
should also be returned with the model. Default is TRUE.}

\item{metaNames}{(Optional) Vector with the column names of data that
correspond to metavariables. Default is NULL.}

\item{keepModels}{(Optional) Boolean indicating if the individual models
should be kept. Can get large in size. Default is TRUE.}

\item{varSelPercent}{Numeric in (0,1) indicating (approx) percentage of
features to keep for each tree.}
}
\value{
A list with  entries
\enumerate{
\item varImportanceData: Data.frame for variable importance
information.
\item (Optional) model: List of CART that builds the random forest model.
\item (Optional) varImportancePlot: Variable importance plots.
}
}
\description{
This function creates a random forest model for data with PCs and
meta-variables. This includes proper combination for variable importance.
Recommend to users to use randomForest_CVPC in general and perhaps just
this for a final model
}
\details{
Upcoming: See TEST comment
Drop parts to make smaller memory impact.
https://rstudio-pubs-static.s3.amazonaws.com/381812_7b5343e0f1a7403fba31ff7d45cf6d88.html
}
\examples{
data <- simulatePP()
pcaData <- getPCAData(data=data, repeatedUniqueId='Image',
                      xRange = c(0,1),  yRange = c(0,1), silent=F)
RF1 <- PCRandomForest(data=pcaData[-2])

pcaMeta <- simulateMeta(pcaData)
RF2 <- PCRandomForest(pcaMeta[-2],
                               metaNames=c("randUnif","randBin","rNorm",
                                           "corrUnif","corrBin","corrNorm"))
}
