% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPP.R
\name{plotPP}
\alias{plotPP}
\title{Plot Spatial Point Process}
\usage{
plotPP(
  data,
  colorGuide = NULL,
  ptSize = 1,
  xlim = c(min(data[, 1]), max(data[, 1])),
  ylim = c(min(data[, 2]), max(data[, 2])),
  dropAxes = F,
  colors = NULL
)
}
\arguments{
\item{data}{Data.frame with x,y, and agent type (in that order)}

\item{colorGuide}{(Optional) String for guides(color=) in ggplot2. Usually
NULL or 'none' is sufficient. Default is NULL.}

\item{ptSize}{(Optional) Numeric indicating point size. Default is 1.}

\item{xlim}{(Optional) Two value numeric vector indicating the size of the
region in the x-direction. Default is c(min(x),max(x)).}

\item{ylim}{(Optional) Two value numeric vector indicating the size of the
region in the y-direction. Default is c(min(y),max(y)).}

\item{dropAxes}{(Optional) Boolean indicating if the x, y axis title and
labels should be dropped. Default is FALSE.}

\item{colors}{(Optional) Vector of colors for the points. Default is NULL, or
ggplot2 selected colors.}
}
\value{
ggplot2 plot of the spatial point process.
}
\description{
This function is used to plot a spatial point process.
}
\examples{
dat <- simulatePP(cellVarData=
                       data.frame('stage'=c(0,1),
                                  'A'=c(0,0),
                                  'B'=c(1/50,1/50)),
                   cellKappaData=data.frame(
                                  'cell'=c('A','B'),
                                  'clusterCell'=c(NA,'A'),
                                  'kappa'=c(20,5)),
                   peoplePerStage=100,
                   imagesPerPerson=1,
                   reduceEdge=0.025,
                   silent=F )
plotPP(dat[dat$Image==1,c('x','y','cellType')])
}
