% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_PC.R
\name{predict.RandomForest_PC}
\alias{predict.RandomForest_PC}
\title{Predict Using RandomForest_PC}
\usage{
\method{predict}{RandomForest_PC}(model, data_pred, type = "all", data = NULL)
}
\arguments{
\item{model}{RandomForest_PC model. See computeRandomForest_PC. A list of
CART models from rpart.}

\item{data_pred}{data.frame of the data to be predicted.}

\item{type}{(Optional) String indicating type of analysis. Options are pred
or all. The choice changes the return to best fit intended use.}

\item{data}{(Optional) data.frame of full data. The data used to fit the
model will be extracted (by row name) to determine data the model knows.}
}
\value{
The returned data depends on type:
1. type='pred': returns a vector of the predictions
2. type='all': returns a vector of the predictions
}
\description{
This function gets the predicted value from a RandomForest_PC
model.
}
\examples{
data_pp <- simulatePP(cellVarData=
                      data.frame('stage'=c(0,1),
                                 'A'=c(0,0),
                                 'B'=c(1/50,1/50)),
                    cellKappaData=data.frame(
                                 'cell'=c('A','B'),
                                 'clusterCell'=c(NA,'A'),
                                 'kappa'=c(20,5)),
                    peoplePerStage=100,
                    imagesPerPerson=1,
                    reduceEdge=0.025,
                    silent=F )
pcaData <- getPCAData(data_pp,repeatedUniqueId='Image',
                      xRange = c(0,1),  yRange = c(0,1), silent=F)
RF <- computeRandomForest_PC(data=pcaData[-2], nTrees = 5)
pred <- predict.RandomForest_PC(model = RF[[1]], type='all',
                                 data_pred = pcaData[-2],
                                 data=pcaData[-2])
}
