% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TNBC.R
\docType{data}
\name{TNBC}
\alias{TNBC}
\title{Triple Negative Breast Cancer Data}
\format{
\subsection{\code{TNBC}}{

A data frame with 33 rows and 1398 columns:
\describe{
\item{Class}{Outcome of each patient}
\item{Person}{Person for each image}
\item{NA_Si_PC1 through tumerYN_tumerYN_PC3}{Principle components of the K
functions for the named interactions}
\item{age}{Meta-variable for patient age}
...
}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
TNBC
}
\description{
A funky model ready set of principle components from K functions based on
triple negative breast cancer data from patients. The original data was
proteins as coded in T/F values. Additionally, the age meta-variable was
added.
}
\keyword{datasets}
