% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TNBC.R
\docType{data}
\name{TNBC_Meta}
\alias{TNBC_Meta}
\title{Triple Negative Breast Cancer Meta Data}
\format{
\subsection{\code{TNBC_Meta}}{

A data frame with 33 rows and 2 columns:
\describe{
\item{Person}{Person for each image}
\item{Age}{Meta-variable for patient age}
...
}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
TNBC_Meta
}
\description{
Data of triple negative breast cancer biopsies from patients, with. Additionally, the age meta-variable was
added.
}
\keyword{datasets}
