% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TNBC.R
\docType{data}
\name{TNBC_pheno}
\alias{TNBC_pheno}
\title{Triple Negative Breast Cancer Data - Phenotypes}
\format{
\subsection{\code{TNBC}}{

A data frame with 170,171 rows and 5 columns:
\describe{
\item{Class}{Outcome of each patient}
\item{Person}{Person for each image}
\item{cellx, celly}{x and y location for each cell}
\item{Phenotype}{Categories for cell phenotype}
...
}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
TNBC_pheno
}
\description{
A subset of triple negative breast cancer data from patients indicating
cell information for each person. The phenotypes are given.
}
\keyword{datasets}
