% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{.generateInvClusterPatterns}
\alias{.generateInvClusterPatterns}
\title{Generate Self/Invisible Clustering Patterns}
\usage{
.generateInvClusterPatterns(
  stageName,
  reduceEdge,
  peoplePerStage,
  imagesPerPerson,
  kappas,
  cellTypes,
  cellVars,
  kappaSep = F,
  imageAdj = 0,
  personAdj = 0
)
}
\arguments{
\item{stageName}{String inidicating the outcome name that should be given.}

\item{reduceEdge}{Numeric indicating how far from edge agents
should be placed. Note the current figure size is \link{0,1}x\link{0,1}.}

\item{peoplePerStage}{Numeric indicating the number of units per
outcome.}

\item{imagesPerPerson}{Numeric indicating the number of repeated
measures.}

\item{kappas}{Numeric vector directing the number of agents per type per
image.}

\item{cellTypes}{Vector with the agent types of interest being generated.
Corresponds to the kappas.}

\item{cellVars}{Vector with the variables for each agent type being
generated. Corresponds to the kappas and cellTypes.}

\item{kappaSep}{(Optional) Boolean. Current use understood but not used. May
remove.}

\item{imageAdj}{(Optional) Numeric to add to image count. Used when images
have already been generated in the resulting dataset and this will be
appended.}
}
\value{
Data.frame containing each point in the defined clustered patterns.
Note, this does not include the data that it clustered around, hence
creating self/invisible clustering.

The data.frame has columns for outcome, x coordinate, y coordinate, agent
type, unit, and unique repeated measure id.
}
\description{
This (internal) function creates self or invisible clustering patterns. It
relies on .generateCSRPatterns and .clusterAroundCells, dropping
unnecessary data.
}
\details{
Upcoming: Determine (generally) how many clusters and how big of clusters to
build.
}
\examples{
# See code for simulatePP. This is not an outward function so won't be
#     viewable.
}
