% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_PC.R
\name{.getUnderlyingVariable}
\alias{.getUnderlyingVariable}
\title{Get Underlying Function}
\usage{
.getUnderlyingVariable(names, returnUnique = T)
}
\arguments{
\item{names}{Vector of names to get underlying function on each}
}
\value{
Vector of underlying functions for each in name.
}
\description{
This (internal) function is used to get the underlying functino (i.e. no _PC#
in it). This is used to centralize called from RandomForest_PC.R and
RandomForest_CVPC.R. If there is no PC (i.e. in meta-varaiables), the
meta-variable is returned.
}
\examples{
# See code for .computeRandomForest_PC. This is not an outward function so
#     won't be viewable.
}
