% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCAData.R
\name{.insertMissingRows}
\alias{.insertMissingRows}
\title{Insert Missing Rows Back into Data}
\usage{
.insertMissingRows(data_add, insertRows)
}
\arguments{
\item{data_add}{A data.frame with the PCs.}

\item{insertRows}{Numeric vector indicating the original rows that were
dropped.}
}
\value{
A data.frame with the PCs, now including the NA rows.
}
\description{
This (internal) function inserting missing observations back into PCA data.
The data was dropped in order to allow PCA, but the NA need to be
re-inserted in order to correctly analyze the data.
}
\details{
See .getPCs for use.
}
