% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest_PC.R
\name{.plotImportance}
\alias{.plotImportance}
\title{Plot Variable Importance (Plot Data)}
\usage{
.plotImportance(varImportanceData)
}
\arguments{
\item{varImportanceData}{Data.frame for the variable importance information.}
}
\value{
grid.arrange containing two ggplots.
}
\description{
This (internal) function plots the variable importance for a
computeRandomForest_PC model.
}
\examples{
# See code for .plotVariableImportance. This is not an outward function so won't be
#     viewable.
}
