% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRandomForest.R
\name{.plotVariableImportance}
\alias{.plotVariableImportance}
\title{Plot Variable Importance (Ensure Data)}
\usage{
.plotVariableImportance(varImportanceData = NULL, model = NULL)
}
\arguments{
\item{varImportanceData}{(Optional) Data.frame for the variable importance
information.}

\item{model}{(Optional) Random forest model from PCRandomForest.}
}
\value{
grid.arrange containing two ggplots
}
\description{
This (internal) function organizes the data to plot the variable importance
for a PCRandomForest model. One of the inputs is necessary.
}
\details{
See use in PCRandomForest.

Upcoming: See TEST comment
}
