% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_K_functions}
\alias{plot_K_functions}
\title{Compare K Functions Between Outcomes}
\usage{
plot_K_functions(data, inc.legend = TRUE, inc.noise = FALSE)
}
\arguments{
\item{data}{Data.frame with the columns r, K, Unit, and Outcome. r indicates
the radius of checked K function, K indicates the K function value, Unit
specifies the unique K function, and Outcome indicates the unit outcome.}

\item{inc.legend}{(Optional) Boolean indicating if the legend should be given.
This will also include numbers to indicate if any K functions are missing.
The default is TRUE.}

\item{inc.noise}{(Optional) Boolean indicating if a gray, dashed line should
be included to show what total noise would be like. The default is FALSE.}
}
\value{
ggplot2 object showing the K function with a superimposed average
}
\description{
Compare K Functions Between Outcomes
}
\examples{
# Example 1
tmp <- getKFunction(TNBC_pheno[TNBC_pheno$Class == 0, -1],
  agents = c("Tumor", "Tumor"),
  unit = "Person",
  rCheckVals = seq(0, 50, 1)
)
tmp1 <- getKFunction(TNBC_pheno[TNBC_pheno$Class == 1, -1],
  agents = c("Tumor", "Tumor"),
  unit = "Person",
  rCheckVals = seq(0, 50, 1)
)
tmp_1 <- tidyr::pivot_longer(data = tmp, cols = K1:K18)
tmp1_1 <- tidyr::pivot_longer(data = tmp1, cols = K1:K15)

data_plot <- rbind(
  data.frame(
    "r" = tmp_1$r,
    "K" = tmp_1$value,
    "Unit" = tmp_1$name,
    "Outcome" = "0"
  ),
  data.frame(
    "r" = tmp1_1$r,
    "K" = tmp1_1$value,
    "Unit" = paste0(tmp1_1$name, "_1"),
    "Outcome" = "1"
  )
)

plot_K_functions(data_plot)

# Example 2
tmp <- getKFunction(TNBC_pheno[TNBC_pheno$Class == 0, -1],
  agents = c("Tumor", "B"), unit = "Person",
  rCheckVals = seq(0, 50, 1)
)
tmp1 <- getKFunction(TNBC_pheno[TNBC_pheno$Class == 1, -1],
  agents = c("Tumor", "B"), unit = "Person",
  rCheckVals = seq(0, 50, 1)
)

tmp_1 <- tidyr::pivot_longer(data = tmp, cols = K1:K18)
tmp1_1 <- tidyr::pivot_longer(data = tmp1, cols = K1:K15)

data_plot <- rbind(
  data.frame(
    "r" = tmp_1$r,
    "K" = tmp_1$value,
    "Unit" = tmp_1$name,
    "Outcome" = "0"
  ),
  data.frame(
    "r" = tmp1_1$r,
    "K" = tmp1_1$value,
    "Unit" = paste0(tmp1_1$name, "_1"),
    "Outcome" = "1"
  )
)

plot_K_functions(data_plot)
}
