% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPP.R
\name{plotPP}
\alias{plotPP}
\title{Plot Spatial Point Process}
\usage{
plotPP(
  data,
  colorGuide = NULL,
  ptSize = 1,
  xlim = c(min(data[, 1]), max(data[, 1])),
  ylim = c(min(data[, 2]), max(data[, 2])),
  dropAxes = FALSE,
  colors = NULL
)
}
\arguments{
\item{data}{Data.frame with x,y, and agent type (in that order)}

\item{colorGuide}{(Optional) String for guides(color=) in ggplot2. Usually
NULL or 'none' is sufficient. Default is NULL.}

\item{ptSize}{(Optional) Numeric indicating point size. Default is 1.}

\item{xlim}{(Optional) Two value numeric vector indicating the size of the
region in the x-direction. Default is c(min(x),max(x)).}

\item{ylim}{(Optional) Two value numeric vector indicating the size of the
region in the y-direction. Default is c(min(y),max(y)).}

\item{dropAxes}{(Optional) Boolean indicating if the x, y axis title and
labels should be dropped. Default is FALSE.}

\item{colors}{(Optional) Vector of colors for the points. Default is NULL, or
ggplot2 selected colors.}
}
\value{
ggplot2 plot of the spatial point process.
}
\description{
This function is used to plot a spatial point process.
}
\examples{
plotPP(TNBC_pheno[TNBC_pheno$Person==1,c("cellx",'celly','Phenotype')])
plotPP(diabetes[diabetes$Image=='E37',c('x','y','cellType')])
}
