% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TNBC.R
\docType{data}
\name{TNBC_Meta}
\alias{TNBC_Meta}
\title{Triple Negative Breast Cancer Data}
\format{
\subsection{\code{TNBC_Meta}}{

A data frame with 33 rows and 2 columns:
\describe{
\item{Person}{Person which the meta-data is related to}
\item{Age}{Age for the related person}
...
}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
TNBC_Meta
}
\description{
A subset of triple negative breast cancer data containing the meta-data
for each person.
}
\keyword{datasets}
