% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TNBC.R
\docType{data}
\name{TNBC_pheno}
\alias{TNBC_pheno}
\title{Triple Negative Breast Cancer Phenotypes}
\format{
\subsection{\code{TNBC_pheno}}{

A data frame with 170,171 rows and 5 columns:
\describe{
\item{Class}{Outcome of each patient}
\item{Person}{Person for which each cell is related}
\item{cellx, celly}{The x-y coordinates of the cell}
\item{Phenotype}{The classified phenotype for the cecll}
...
}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
TNBC_pheno
}
\description{
Data of triple negative breast cancer biopsies from patients.
}
\keyword{datasets}
